requirements.txt
# Core Dependencies
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3

# Computer Vision & AI
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
ultralytics==8.0.20  # YOLO v8
torch==2.0.1
torchvision==0.15.2
tensorflow-lite==2.13.0

# Hardware Communication
pyserial==3.5
pymavlink==2.4.37
pynmea2==1.19.0
RPi.GPIO==0.7.1  # Raspberry Pi GPIO

# Networking & Communication
websockets==11.0.3
requests==2.31.0
paho-mqtt==1.6.1
gevent==23.7.0
eventlet==0.33.3

# Database
sqlite3  # Built-in with Python
psycopg2-binary==2.9.7  # PostgreSQL adapter

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
watchdog==3.0.0
schedule==1.2.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
black==23.7.0
flake8==6.0.0
pre-commit==3.4.0

# Geospatial
geopy==2.4.0
folium==0.14.0

# Image Processing
scikit-image==0.21.0
imageio==2.31.3

# Data Analysis
pandas==2.0.3
matplotlib==3.7.2

# Async Support
asyncio
aiofiles==23.2.1

# Security
cryptography==41.0.4
bcrypt==4.0.1

# Logging
loguru==0.7.0

# System Monitoring
psutil==5.9.5
